+++ src/MongooseHttpServer.h
@@ -13,382 +13,25 @@
 #include <mongoose.h>
 
 #include <functional>
+#include <list>
 
 #include "MongooseString.h"
-#include "MongooseHttp.h"
-#include "MongooseSocket.h"
+#include "MongooseHttpServerConnection.h"
+#include "MongooseHttpServerEndpoint.h"
+#include "MongooseHttpServerEndpointUpload.h"
+#include "MongooseHttpServerEndpointWebsocket.h"
 
-// Make a copy of the HTTP header so it is avalible outside of the onReceive
-// callback. Setting to 0 will save some runtime memory but accessing the HTTP
-// message details outside of the onReceive callback will give undefined behaviour.
-// The body may not allways be avalible even in onReceive, eg file upload
-#ifndef MG_COPY_HTTP_MESSAGE
-#define MG_COPY_HTTP_MESSAGE 1
-#endif
 
-class MongooseHttpServer;
-class MongooseHttpServerRequest;
-class MongooseHttpServerResponse;
-class MongooseHttpServerResponseBasic;
-#ifdef ARDUINO
-class MongooseHttpServerResponseStream;
-#endif
-#if MG_ENABLE_HTTP_WEBSOCKET
-class MongooseHttpWebSocketConnection;
-#define MG_F_IS_MongooseHttpWebSocketConnection MG_F_USER_1
-#endif
-class MongooseHttpServerEndpoint;
-
-class MongooseHttpServerRequest
+class MongooseHttpServer : public MongooseHttpServerConnection
 {
-  friend MongooseHttpServerEndpoint;
 
+  private:
+    std::list<MongooseHttpServerEndpoint *> _endpoints;
+    MongooseHttpServerEndpoint _notFound;
+
+    HttpRequestMethodComposite method(mg_str method);
   protected:
-    mg_connection *_nc;
-    http_message *_msg;
-    HttpRequestMethodComposite _method;
-    MongooseHttpServerResponse *_response;
-    bool _responseSent;
-
-    void sendBody();
-
-#if MG_COPY_HTTP_MESSAGE
-    http_message *duplicateMessage(http_message *);
-#endif
-
-  public:
-    MongooseHttpServerRequest(mg_connection *nc, http_message *msg);
-    virtual ~MongooseHttpServerRequest();
-
-    virtual bool isUpload() { return false; }
-    virtual bool isWebSocket() { return false; }
-
-    HttpRequestMethodComposite method() {
-      return _method;
-    }
-
-    MongooseString message() {
-      return MongooseString(_msg->message);
-    }
-    MongooseString body() {
-      return MongooseString(_msg->body);
-    }
-
-    MongooseString methodStr() {
-      return MongooseString(_msg->method);
-    }
-    MongooseString uri() {
-      return MongooseString(_msg->uri);
-    }
-    MongooseString proto() {
-      return MongooseString(_msg->proto);
-    }
-
-    int respCode() {
-      return _msg->resp_code;
-    }
-    MongooseString respStatusMsg() {
-      return MongooseString(_msg->resp_status_msg);
-    }
-
-    MongooseString queryString() {
-      return MongooseString(_msg->query_string);
-    }
-
-    int headers() {
-      int i;
-      for (i = 0; i < MG_MAX_HTTP_HEADERS && _msg->header_names[i].len > 0; i++) {
-      }
-      return i;
-    }
-    MongooseString headers(const char *name) {
-      MongooseString ret(mg_get_http_header(_msg, name));
-      return ret;
-    }
-    MongooseString headerNames(int i) {
-      return MongooseString(_msg->header_names[i]);
-    }
-    MongooseString headerValues(int i) {
-      return MongooseString(_msg->header_values[i]);
-    }
-
-    MongooseString host() {
-      return headers("Host");
-    }
-
-    MongooseString contentType() {
-      return headers("Content-Type");
-    }
-
-    size_t contentLength() {
-      return _msg->body.len;
-    }
-
-    void redirect(const char *url);
-#ifdef ARDUINO
-    void redirect(const String& url);
-#endif
-
-    MongooseHttpServerResponseBasic *beginResponse();
-
-#ifdef ARDUINO
-    MongooseHttpServerResponseStream *beginResponseStream();
-#endif
-
-    // Takes ownership of `response`, will delete when finished. Do not use `response` after calling
-    void send(MongooseHttpServerResponse *response);
-    bool responseSent() {
-      return NULL != _response;
-    }
-
-    void send(int code);
-    void send(int code, const char *contentType, const char *content="");
-#ifdef ARDUINO
-    void send(int code, const String& contentType, const String& content=String());
-#endif
-
-    bool hasParam(const char *name) const;
-#ifdef ARDUINO
-    bool hasParam(const String& name) const;
-    bool hasParam(const __FlashStringHelper * data) const;
-#endif
-
-    int getParam(const char *name, char *dst, size_t dst_len) const;
-#ifdef ARDUINO
-    int getParam(const String& name, char *dst, size_t dst_len) const;
-    int getParam(const __FlashStringHelper * data, char *dst, size_t dst_len) const;
-#endif
-
-#ifdef ARDUINO
-    String getParam(const char *name) const;
-    String getParam(const String& name) const;
-    String getParam(const __FlashStringHelper * data) const;
-#endif
-
-    bool authenticate(const char * username, const char * password);
-#ifdef ARDUINO
-    bool authenticate(const String& username, const String& password) {
-      return authenticate(username.c_str(), password.c_str());
-    }
-#endif
-    void requestAuthentication(const char* realm);
-#ifdef ARDUINO
-    void requestAuthentication(const String& realm) {
-      requestAuthentication(realm.c_str());
-    }
-#endif
-};
-
-class MongooseHttpServerRequestUpload : public MongooseHttpServerRequest
-{
-  friend MongooseHttpServerEndpoint;
-
-  private:
-    uint64_t index;
-
-  public:
-    MongooseHttpServerRequestUpload(mg_connection *nc, http_message *msg) :
-      MongooseHttpServerRequest(nc, msg),
-      index(0)
-    {
-    }
-    virtual ~MongooseHttpServerRequestUpload() {
-    }
-
-    virtual bool isUpload() { return true; }
-};
-
-class MongooseHttpServerResponse
-{
-  private:
-    int _code;
-    char *_contentType;
-    int64_t _contentLength;
-
-    char * _headerBuffer;
-
-  public:
-    MongooseHttpServerResponse();
-    virtual ~MongooseHttpServerResponse();
-
-    void setCode(int code) {
-      _code = code;
-    }
-    void setContentType(const char *contentType);
-    void setContentLength(int64_t contentLength) {
-      _contentLength = contentLength;
-    }
-
-    bool addHeader(const char *name, const char *value);
-#ifdef ARDUINO
-    void setContentType(String &contentType) {
-      setContentType(contentType.c_str());
-    }
-    void setContentType(const __FlashStringHelper *contentType);
-    bool addHeader(const String& name, const String& value);
-#endif
-
-    // send the to `nc`, return true if more to send
-    virtual void sendHeaders(struct mg_connection *nc);
-
-    // send (a part of) the body to `nc`, return < `bytes` if no more to send
-    virtual size_t sendBody(struct mg_connection *nc, size_t bytes) = 0;
-};
-
-class MongooseHttpServerResponseBasic:
-  public MongooseHttpServerResponse
-{
-  private:
-    const uint8_t *ptr;
-    size_t len;
-
-  public:
-    MongooseHttpServerResponseBasic();
-
-    void setContent(const char *content);
-    void setContent(const uint8_t *content, size_t len);
-    void setContent(MongooseString &content) {
-      setContent((const uint8_t *)content.c_str(), content.length());
-    }
-    virtual size_t sendBody(struct mg_connection *nc, size_t bytes);
-};
-
-#ifdef ARDUINO
-class MongooseHttpServerResponseStream:
-  public MongooseHttpServerResponse,
-  public Print
-{
-  private:
-    mbuf _content;
-
-  public:
-    MongooseHttpServerResponseStream();
-    virtual ~MongooseHttpServerResponseStream();
-
-    size_t write(const uint8_t *data, size_t len);
-    size_t write(uint8_t data);
-  //  using Print::write;
-
-    virtual size_t sendBody(struct mg_connection *nc, size_t bytes);
-};
-#endif
-
-
-typedef std::function<void(MongooseHttpServerRequest *request)> MongooseHttpRequestHandler;
-typedef std::function<size_t(MongooseHttpServerRequest *request, int ev, MongooseString filename, uint64_t index, uint8_t *data, size_t len)> MongooseHttpUploadHandler;
-#if MG_ENABLE_HTTP_WEBSOCKET
-typedef std::function<void(MongooseHttpWebSocketConnection *connection)> MongooseHttpWebSocketConnectionHandler;
-typedef std::function<void(MongooseHttpWebSocketConnection *connection, int flags, uint8_t *data, size_t len)> MongooseHttpWebSocketFrameHandler;
-#endif
-
-class MongooseHttpServerEndpoint : public MongooseSocket
-{
-  private:
-    HttpRequestMethodComposite _method;
-    MongooseHttpRequestHandler _request;
-    MongooseHttpUploadHandler _upload;
-    MongooseHttpRequestHandler _close;
-#if MG_ENABLE_HTTP_WEBSOCKET
-    MongooseHttpWebSocketConnectionHandler _wsConnect;
-    MongooseHttpWebSocketFrameHandler _wsFrame;
-#endif
-
-    void onEvent(mg_connection *nc, int ev, void *p);
-    void onPoll(mg_connection *nc);
-    void onSend(mg_connection *nc, int num_bytes) {
-      onPoll(nc);
-    }
-    void onClose(mg_connection *nc);
-
-  public:
-    MongooseHttpServerEndpoint(HttpRequestMethodComposite method) :
-      MongooseSocket(),
-      _method(HTTP_ANY),
-      _request(NULL),
-      _upload(NULL)
-#if MG_ENABLE_HTTP_WEBSOCKET
-      ,
-      _wsConnect(NULL),
-      _wsFrame(NULL)
-#endif
-    {
-
-    }
-
-    bool attach(mg_connection *nc, const char *uri = nullptr, HttpRequestMethodComposite method = HTTP_ANY)
-    {
-      if(uri) {
-        _method = method;
-        mg_register_http_endpoint(nc, uri, eventHandler, this);
-      } else {
-        mg_set_protocol_http_websocket(nc);
-      }
-
-      return true;
-    }
-
-    MongooseHttpServerEndpoint *onRequest(MongooseHttpRequestHandler handler) {
-      this->_request = handler;
-      return this;
-    }
-
-    MongooseHttpServerEndpoint *onUpload(MongooseHttpUploadHandler handler) {
-      this->_upload = handler;
-      return this;
-    }
-
-    MongooseHttpServerEndpoint *onClose(MongooseHttpRequestHandler handler) {
-      this->_close = handler;
-      return this;
-    }
-
-#if MG_ENABLE_HTTP_WEBSOCKET
-    MongooseHttpServerEndpoint *onConnect(MongooseHttpWebSocketConnectionHandler handler) {
-      this->_wsConnect = handler;
-      return this;
-    }
-
-    MongooseHttpServerEndpoint *onFrame(MongooseHttpWebSocketFrameHandler handler) {
-      this->_wsFrame = handler;
-      return this;
-    }
-#endif
-};
-
-#if MG_ENABLE_HTTP_WEBSOCKET
-class MongooseHttpWebSocketConnection : public MongooseHttpServerRequest
-{
-  friend MongooseHttpServerEndpoint;
-
-  public:
-    MongooseHttpWebSocketConnection(mg_connection *nc, http_message *msg);
-    virtual ~MongooseHttpWebSocketConnection();
-
-    virtual bool isWebSocket() { return true; }
-
-    void send(int op, const void *data, size_t len);
-    void send(const char *buf) {
-      send(WEBSOCKET_OP_TEXT, buf, strlen(buf));
-    }
-#ifdef ARDUINO
-    void send(String &str) {
-      send(str.c_str());
-    }
-#endif
-
-    const union socket_address *getRemoteAddress() {
-      return &(_nc->sa);
-    }
-    const mg_connection *getConnection() {
-      return _nc;
-    }
-};
-#endif
-
-class MongooseHttpServer : public MongooseHttpServerEndpoint
-{
-  protected:
-    void onEvent(int ev, void *p);
+    void onHeaders(mg_connection *nc, mg_http_message *msg);
 
   public:
     MongooseHttpServer();
@@ -401,12 +44,14 @@ class MongooseHttpServer : public MongooseHttpServerEndpoint
     MongooseHttpServerEndpoint *on(const char* uri, HttpRequestMethodComposite method);
     MongooseHttpServerEndpoint *on(const char* uri, MongooseHttpRequestHandler onRequest);
     MongooseHttpServerEndpoint *on(const char* uri, HttpRequestMethodComposite method, MongooseHttpRequestHandler onRequest);
+    MongooseHttpServerEndpointUpload *on(const char* uri, MongooseHttpUploadHandler onUpload);
+    MongooseHttpServerEndpointWebSocket *on(const char* uri, MongooseHttpWebSocketFrameHandler onFrame);
+    MongooseHttpServerEndpoint *on(MongooseHttpServerEndpoint *endpoint);
 
     void onNotFound(MongooseHttpRequestHandler fn);
 
     void reset();
 
-#if MG_ENABLE_HTTP_WEBSOCKET
     void sendAll(MongooseHttpWebSocketConnection *from, const char *endpoint, int op, const void *data, size_t len);
 
     void sendAll(MongooseHttpWebSocketConnection *from, int op, const void *data, size_t len) {
@@ -444,7 +89,6 @@ class MongooseHttpServer : public MongooseHttpServerEndpoint
       sendAll(endpoint, str.c_str());
     }
 #endif
-#endif // MG_ENABLE_HTTP_WEBSOCKET
 };
 
 #endif /* _MongooseHttpServer_H_ */
