+++ src/MongooseHttpServerConnection.h
@@ -1,0 +1,43 @@
+#ifndef MongooseHttpServerConnection_h
+#define MongooseHttpServerConnection_h
+
+#ifdef ARDUINO
+#include "Arduino.h"
+
+#if defined(ESP8266)
+#include <ESP8266WiFi.h>
+#endif
+
+#endif
+
+#include <mongoose.h>
+
+#include <functional>
+
+#include "MongooseString.h"
+#include "MongooseHttp.h"
+#include "MongooseSocket.h"
+
+class MongooseHttpServerConnection : public MongooseSocket
+{
+  private:
+  
+  protected:
+    void onEvent(mg_connection *nc, int ev, void *p);
+
+    virtual void onHeaders(mg_connection *nc, mg_http_message *msg) { }
+    virtual void onMessage(mg_connection *nc, mg_http_message *msg) { }
+    virtual void onWebSocketConnect(mg_connection *nc, mg_http_message *msg) { }
+    virtual void onWebSocketMessage(mg_connection *nc, mg_ws_message *msg) { }
+    virtual void onWebSocketControl(mg_connection *nc, mg_ws_message *msg) { }
+
+  public:
+    MongooseHttpServerConnection() :
+      MongooseSocket()
+    {
+    }
+    virtual ~MongooseHttpServerConnection() {
+    }
+};
+
+#endif /* MongooseHttpServerConnection_h */
